Register both the encoded and unencoded versions of the vision_plural in URLs
Sometimes we use the unencoded version directly, for example in links to individual
visions from the home page. Though the browser will encode the URL once it's in the
location bar, it will match against the exact text of the link -- i.e., the unencoded
version.

These unencoded uses of vision_plural in URLs should be considered deprecated and
should be phased out.